refactor: (#19) useRoom UI 로직과 비즈니스 로직 분리
